Cache Matrix Program
This program is designed to demonstrate the concept of Lexical Scoping
in R. In the program, there is a pair of functions: makeCacheMatrix and
cacheSolve.
makeCacheMatrix stores a matrix input, or calculates the inverse of the
matrix.
cacheSolve uses the inverse of the matrix from makeCacheMatrix when it
exists, or calculates the inverse of the matrix if it was not solved.